Don't bother with nesting in prefab folder
This was a relic from back when prefabs were added to the storage folder
by the user. They were allowed to nest them in folders to make
organization easier, but now that prefabs are managed by the plugin
this is no longer needed.

If in the future the user goes need some form of organization, we can
introduce a tagging system with the insert menu so that they can quickly
filter between props, buildings, etc.